feat: update user and subscription handling with new database schema
- Added support for the `@mistralai/mistralai` package in the project dependencies.
- Refactored user-related services to utilize the new database schema, including changes to user and subscription management.
- Removed deprecated repository classes and integrated direct database queries for user and subscription operations.
- Updated various components and services to align with the new data structures, enhancing overall code efficiency and maintainability.
