+++ cook/populate-github-user.ts
@@ -1,5 +1,7 @@
-import { UserRepository } from "@/core/repositories/user-repository";
-import { PopulateGithubUser } from "@/core/services/populate-github-user.service";
+import { db } from "@/db";
+import { user as userTable } from "@/db/schema";
+import { PopulateGithubUser } from "@/services/populate-github-user.service";
+import { eq } from "drizzle-orm";
 
 if (require.main === module) {
 	(async () => {
@@ -22,8 +24,9 @@ if (require.main === module) {
 				twitterUsername = args[2];
 			}
 
-			const userRepository = new UserRepository();
-			const existingUser = await userRepository.findByUsername(username);
+			const existingUser = await db.query.user.findFirst({
+				where: eq(userTable.username, username),
+			});
 			if (existingUser) {
 				console.log(
 					`github username ${username} already processed with id ${existingUser.id}, skipping`,
