+++ drizzle/0006_add-pinned-repos-to-user.sql
@@ -1,5 +1,1 @@
--- ✨ UP
 ALTER TABLE "user" ADD COLUMN "pinned_repos" jsonb[] DEFAULT '{}'::jsonb[];
-
--- 🔄 DOWN
-ALTER TABLE "user" DROP COLUMN "pinned_repos";
