--- drizzle/0009_add_curriculum_vitae_into_users.sql
@@ -1,5 +1,0 @@
--- ✨ UP
-ALTER TABLE "user" ADD COLUMN "curriculum_vitae" jsonb DEFAULT '{}'::jsonb NOT NULL;--> statement-breakpoint
-
--- 🔄 DOWN
-ALTER TABLE "user" DROP COLUMN "curriculum_vitae";
