+++ drizzle/0010_certain_forgotten_one.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "user" ALTER COLUMN "curriculum_vitae" DROP DEFAULT;
