+++ drizzle/meta/0010_snapshot.json
@@ -1,6 +1,6 @@
 {
-	"id": "51073147-20fb-497b-8bea-c63e6fddd6ee",
-	"prevId": "55fd8ec1-4f22-4c46-8bc6-cbd5e77a97e1",
+	"id": "c514e29e-e2d8-4795-9ca3-57435a41ac89",
+	"prevId": "51073147-20fb-497b-8bea-c63e6fddd6ee",
 	"version": "7",
 	"dialect": "postgresql",
 	"tables": {
@@ -161,8 +161,7 @@
 					"name": "curriculum_vitae",
 					"type": "jsonb",
 					"primaryKey": false,
-					"notNull": false,
-					"default": "'{}'::jsonb"
+					"notNull": false
 				}
 			},
 			"indexes": {
