--- src/core/models/user-subscription.ts
@@ -1,14 +1,0 @@
-import { z } from "zod";
-
-export const UserSubscriptionSchema = z.object({
-	id: z.string(),
-
-	polarCustomerId: z.string(),
-	polarProductId: z.string(),
-
-	userId: z.string(),
-	subscriptionPlanId: z.string(),
-	active: z.boolean(),
-});
-
-export type UserSubscription = z.infer<typeof UserSubscriptionSchema>;
