+++ src/db/schema/user/pinned-repo.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+export const PinnedRepoSchema = z.object({
+	fullName: z.string(),
+	description: z.string().nullable(),
+	stars: z.number(),
+});
+
+export type PinnedRepo = z.infer<typeof PinnedRepoSchema>;
