+++ src/triggers/get-user-metadata.ts
@@ -1,4 +1,4 @@
-import { RepoOfUserSchema } from "@/core/models/user";
+import { RecentRepoSchema } from "@/db/schema";
 import { UserMetadataExtractor } from "@/services/user-metadata-extractor";
 import { schemaTask } from "@trigger.dev/sdk/v3";
 import { z } from "zod";
@@ -7,7 +7,7 @@ export const getUserMetadata = schemaTask({
 	id: "get-user-metadata",
 	schema: z.object({
 		username: z.string(),
-		repos: RepoOfUserSchema.array(),
+		repos: RecentRepoSchema.array(),
 	}),
 	run: async ({ username, repos }) => {
 		return new UserMetadataExtractor().extract(username, repos);
